css/css-logical-properties-1/logicalprops-margin-inline.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Logical Properties: margin-inline{-start,end}</title>
<link rel="author" title="Xu Xing" href="mailto:openxu@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-logical-props-1/#logical-prop">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#logical-to-physical">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/style-check.js"></script>

<style>
div {
  border: 1px solid #000;
}

/*Horizontal-tb mode*/
#div_htb_ltr {
  margin-inline-start: 100px;
  margin-inline-end: 80px;
}
#ref_div_htb_ltr {
  margin-left: 100px;
  margin-right: 80px;
}

#div_htb_rtl {
  direction: rtl;
  margin-inline-start: 100px;
  margin-inline-end: 80px;
}
#ref_div_htb_rtl {
  direction: rtl;
  margin-right: 100px;
  margin-left: 80px;
}

#div_htb_override {
  margin-inline-start: 100px;
  margin-inline-end: 80px;
  margin-left: 110px;
  margin-right: 90px;
}
#ref_div_htb_override {
  margin-left: 110px;
  margin-right: 90px;
}

#div_htb_override2 {
  margin-left: 110px;
  margin-right: 90px;
  margin-inline-start: 100px;
  margin-inline-end: 80px;
}
#ref_div_htb_override2 {
  margin-left: 100px;
  margin-right: 80px;
}

/*Vertical-lr mode, the same as vertical-rl mode. Direction is ltr*/
#div_vlr_ltr {
  writing-mode: vertical-lr;
  margin-inline-start: 100px;
  margin-inline-end: 80px;
}
#ref_div_vlr_ltr {
  writing-mode: vertical-lr;
  margin-top: 100px;
  margin-bottom: 80px;
}

#div_vlr_ltr_override {
  writing-mode: vertical-lr;
  margin-inline-start: 100px;
  margin-inline-end: 80px;
  margin-top: 110px;
  margin-bottom: 90px;
}
#ref_div_vlr_ltr_override {
  writing-mode: vertical-lr;
  margin-top: 110px;
  margin-bottom: 90px;
}

#div_vlr_ltr_override2 {
  writing-mode: vertical-lr;
  margin-top: 110px;
  margin-bottom: 90px;
  margin-inline-start: 100px;
  margin-inline-end: 80px;
}
#ref_div_vlr_ltr_override2 {
  writing-mode: vertical-lr;
  margin-top: 100px;
  margin-bottom: 80px;
}

/*Vertical-lr mode, the same as vertical-rl mode. Direction is rtl*/
#div_vlr_rtl {
  writing-mode: vertical-lr;
  direction: rtl;
  margin-inline-start: 100px;
  margin-inline-end: 80px;
}
#ref_div_vlr_rtl {
  writing-mode: vertical-lr;
  direction: rtl;
  margin-bottom: 100px;
  margin-top: 80px;
}

#div_vlr_rtl_override {
  writing-mode: vertical-lr;
  direction: rtl;
  margin-inline-start: 100px;
  margin-inline-end: 80px;
  margin-bottom: 110px;
  margin-top: 90px;
}
#ref_div_vlr_rtl_override {
  writing-mode: vertical-lr;
  direction: rtl;
  margin-bottom: 110px;
  margin-top: 90px;
}

#div_vlr_rtl_override2 {
  writing-mode: vertical-lr;
  direction: rtl;
  margin-bottom: 110px;
  margin-top: 90px;
  margin-inline-start: 100px;
  margin-inline-end: 80px;
}
#ref_div_vlr_rtl_override2 {
  writing-mode: vertical-lr;
  direction: rtl;
  margin-bottom: 100px;
  margin-top: 80px;
}
</style>

<div id="div_htb_ltr"></div>
<div id="ref_div_htb_ltr"></div>
<div id="div_htb_rtl"></div>
<div id="ref_div_htb_rtl"></div>
<div id="div_htb_override"></div>
<div id="ref_div_htb_override"></div>
<div id="div_htb_override2"></div>
<div id="ref_div_htb_override2"></div>

<div id="div_vlr_ltr"></div>
<div id="ref_div_vlr_ltr"></div>
<div id="div_vlr_ltr_override">1</div>
<div id="ref_div_vlr_ltr_override">1</div>
<div id="div_vlr_ltr_override2"></div>
<div id="ref_div_vlr_ltr_override2"></div>

<div id="div_vlr_rtl"></div>
<div id="ref_div_vlr_rtl"></div>
<div id="div_vlr_rtl_override"></div>
<div id="ref_div_vlr_rtl_override"></div>
<div id="div_vlr_rtl_override2"></div>
<div id="ref_div_vlr_rtl_override2"></div>

<script>
test(function () {
  assert_true(compareInlineMargin_htb("div_htb_ltr", "ref_div_htb_ltr"));
  assert_true(compareInlineMargin_htb("div_htb_rtl", "ref_div_htb_rtl"));
  assert_true(compareInlineMargin_htb("div_htb_override", "ref_div_htb_override"));
  assert_true(compareInlineMargin_htb("div_htb_override2", "ref_div_htb_override2"));

  assert_true(compareInlineMargin_vertical("div_vlr_ltr", "ref_div_vlr_ltr"));
  assert_true(compareInlineMargin_vertical("div_vlr_ltr_override", "ref_div_vlr_ltr_override"));
  assert_true(compareInlineMargin_vertical("div_vlr_ltr_override2", "ref_div_vlr_ltr_override2"));

  assert_true(compareInlineMargin_vertical("div_vlr_rtl", "ref_div_vlr_rtl"));
  assert_true(compareInlineMargin_vertical("div_vlr_rtl_override", "ref_div_vlr_rtl_override"));
  assert_true(compareInlineMargin_vertical("div_vlr_rtl_override2", "ref_div_vlr_rtl_override2"));
}, "Check that margin-inline{-start,end} in different writing mode");
</script>
